Move common lib to test, use it in FOSS tests
