--- test/unit/__init__.py
@@ -1,24 +1,0 @@
-# Copyright 2023 Ericsson AB
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-"""
-Setup module paths and environment variables for the functional tests.
-"""
-
-import os
-import sys
-
-# Allow relative imports within the test project to work as expected
-# Without it no module (test) would be able to import the common library
-sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__))))
